implement new MapSlider wrapping Slider
