+++ app/src/main/java/com/kodeco/android/coordplot/app/widgets/Map.kt
@@ -1,0 +1,54 @@
+package com.kodeco.android.coordplot.app.widgets
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.kodeco.android.coordplot.R
+import com.kodeco.android.coordplot.ui.theme.MyApplicationTheme
+
+
+// TODO Move Map() and MapPreview() to their own file
+@Composable
+fun Map(xPercent: Float, yPercent: Float, modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier
+            .size(300.dp)
+            .background(MaterialTheme.colorScheme.tertiary),
+    ) {
+        // TODO Swap this Box out for an Image.
+        //  Make sure to keep the size and offset modifiers.
+        Box(
+            modifier = Modifier
+                .size(36.dp)
+                .offset((xPercent * 300 - 18).dp, (yPercent * 300 - 18).dp)
+                .clip(CircleShape)
+                .background(Color.Blue)
+        )
+        Image(
+            painter = painterResource(id = R.drawable.logo),
+            contentDescription = "Star pin",
+            modifier = Modifier
+                .size(36.dp)
+                .offset((xPercent * 300 - 18).dp, (yPercent * 300 - 18).dp),
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun MapPreview() {
+    MyApplicationTheme {
+        Map(xPercent = 0.5f, yPercent = 0.5f)
+    }
+}
+
