+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <!-- TODO Change this app_name to something more relatable! -->
     <string name="app_name">My Application</string>
-    <!-- TODO Add string resources here for X and Y axis slider titles -->
+    <string name="x_axis">X Axis</string>
+    <string name="y_axis">Y Axis</string>
 </resources>
